Refactor: Consolidate and streamline test fixture imports across test files
